Move custom matchers to their own file; all tests passing
